tests: Make tox notice dependency changes automatically
When passing dependencies via a file using `deps = -rtest-requirements.txt`
as we do in the `tox.ini` file, when we run an upgrade from MetalK8s 2.4 to 2.5
and then invoke fast and slow tests: `Tox` is unable to notice a change in the
requirements file.

It seems pip does not reparse this file and
as such it does not trigger a rebuild of the virtual environment because of the
existing cache version.

A workaround to this was to explicitly define each of the dependencies and stop
using  `-r` but this approach is a dirty fix.

In the PR we resolve to forcing a recreation of the virtual env each time
a test is ran.