salt: fix unsafe approach for etcd `initial-cluster-state
Initially, we set etcd cluster state to new when the member
list from the pillars is empty.

This approach is not optimal since if for what ever reason we are
not able to retrieve the list.

Now, we raise an error when setting the cluster state to `existing`
if we are unable to obtain a list of members.
Also, we only set to `new` state only during a bootstrap.