authentication: add extra check for container status
Initially, we didn't check the status of containers for pods in
nginx-ingress control plane. As such we triggered tests while container
state was unready making most of the tests to fail in the CI.

Now, we ensure to check the status of pods and their corresponding
containers before launching tests.