docs/install: Rework reference architectures
The single architecture that used to be in the Quickstart guide does not
provide sufficient information to readers about the flexibility
available when choosing an architecture for their clusters.
We replace the previous "Installation Plan" section with a few example
architecture and recommendations on how to define one.

References to the previous "quickstart" architecture were removed.

Issue: #2112